+++ R/utils.R
@@ -247,3 +247,39 @@ retry_api_call <- function(expr){
   
   resp
 }
+
+
+# Create table With parameters
+ctas_sql_with <- function(partition = NULL, s3.location = NULL, file.type = "NULL", compress = TRUE){
+  if(file.type!="NULL" || !is.null(s3.location) || !is.null(partition)){
+    FILE <- switch(file.type,
+                   "csv" = "format = 'TEXTFILE',\nfield_delimiter = ','",
+                   "tsv" = "format = 'TEXTFILE',\nfield_delimiter = '\t'",
+                   "parquet" = "format = 'PARQUET'",
+                   "json" = "format = 'JSON'",
+                   "orc" = "format = 'ORC'",
+                   "")
+    
+    COMPRESSION <- ""
+    if (compress) {
+      if(file.type %in% c("tsv", "csv", "json")) warning("Can only compress parquet or orc files: https://docs.aws.amazon.com/athena/latest/ug/create-table-as.html", call. = FALSE)
+      COMPRESSION <- switch(file.type,
+                            "parquet" = ",\nparquet_compression = 'SNAPPY'",
+                            "orc" = ",\norc_compression = 'SNAPPY'",
+                            "")
+    }
+    
+    LOCATION <- if(!is.null(s3.location)){
+      if(file.type == "NULL") paste0("external_location ='", s3.location, "'")
+      else paste0(",\nexternal_location ='", s3.location, "'")
+    } else ""
+    
+    PARTITION <- if(!is.null(partition)){
+      partition <- paste(partition, collapse = "','")
+      if(is.null(s3.location) && file.type == "NULL") paste0("partitioned_by = ARRAY['",partition,"']")
+      else paste0(",\npartitioned_by = ARRAY['",partition,"']")
+    } else ""
+    
+    paste0("WITH (", FILE, COMPRESSION, LOCATION, PARTITION,")\n")
+  } else ""
+}